Refactorgin History Pages on Mobile App (#194)
