+++ backend/GetGo-BE/src/main/resources/application.properties
@@ -2,9 +2,9 @@ spring.application.name=GetGo-BE
 server.port=8080
 
 
-spring.datasource.url=jdbc:postgresql://localhost:5432/database10
+spring.datasource.url=jdbc:postgresql://localhost:5432/baza
 spring.datasource.username=postgres
-spring.datasource.password=MerisPSQL
+spring.datasource.password=postgres
 
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=false
@@ -27,7 +27,9 @@ spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
 
 upload.dir=uploads
+
 app.upload.base-url=http://localhost:8080/
+app.backend.base-url=https://nonpossibly-nonderivable-teddy.ngrok-free.dev
 
 spring.profiles.active=dev
