+++ mobileApp/app/src/main/res/layout/nav_header.xml
@@ -19,7 +19,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/driver_panel"
+        android:text="My Panel"
         android:textColor="@color/white"/>
 
 </LinearLayout>
